refactor: remove pricing strategy classes
