+++ src/chains/config/arbitrum.config.ts
@@ -3,12 +3,11 @@ import { Network } from '@badger-dao/sdk';
 import RPC from '../../config/rpc.config';
 import { TOKENS } from '../../config/tokens.config';
 import { arbitrumTokensConfig } from '../../tokens/config/arbitrum-tokens.config';
-import { BaseStrategy } from '../strategies/base.strategy';
 import { Chain } from './chain.config';
 
 export class Arbitrum extends Chain {
   constructor() {
-    super(Network.Arbitrum, arbitrumTokensConfig, RPC[Network.Arbitrum], new BaseStrategy(Network.Arbitrum));
+    super(Network.Arbitrum, arbitrumTokensConfig, RPC[Network.Arbitrum]);
     Chain.register(this.network, this);
   }
