+++ src/chains/config/polygon.config.ts
@@ -3,12 +3,11 @@ import { Network } from '@badger-dao/sdk';
 import rpc from '../../config/rpc.config';
 import { TOKENS } from '../../config/tokens.config';
 import { maticTokensConfig } from '../../tokens/config/polygon-tokens.config';
-import { BaseStrategy } from '../strategies/base.strategy';
 import { Chain } from './chain.config';
 
 export class Polygon extends Chain {
   constructor() {
-    super(Network.Polygon, maticTokensConfig, rpc[Network.Polygon], new BaseStrategy(Network.Polygon));
+    super(Network.Polygon, maticTokensConfig, rpc[Network.Polygon]);
     Chain.register(this.network, this);
   }
