--- src/chains/strategies/chain.strategy.ts
@@ -1,5 +1,0 @@
-import { TokenPrice } from '../../prices/interface/token-price.interface';
-
-export abstract class ChainStrategy {
-  abstract getPrice(address: string): Promise<TokenPrice>;
-}
